requirements.txt
h5py
skorch
torchvision
matplotlib
tqdm
scikit-learn
joblib
hydra
tensorboard
torch-scatter 
torch-sparse 
torch-cluster 
torch-geometric
seaborn
torchtext
spacy
h5py